Fixed: #395: compile errors and warnings when sizeof(size_t) != sizeof(uint), multiple compiler errors on 64bit Linux
git-svn-id: http://dev.ryzom.com/svn/trunk@1076 60ea55b9-4480-4529-b8ac-ed5291525b0f

+++ nel/src/sound/audio_mixer_user.cpp
@@ -181,12 +181,12 @@ void CAudioMixerUser::initClusteredSound(NL3D::CScene *scene, float minGain, flo
 }
 
 
-void CAudioMixerUser::setPriorityReserve(TSoundPriority priorityChannel, uint reserve)
+void CAudioMixerUser::setPriorityReserve(TSoundPriority priorityChannel, size_t reserve)
 {
 	_PriorityReserve[priorityChannel] = min(_Tracks.size(), reserve);
 }
 
-void CAudioMixerUser::setLowWaterMark(uint value)
+void CAudioMixerUser::setLowWaterMark(size_t value)
 {
 	_LowWaterMark = min(_Tracks.size(), value);
 }
